+++ lib/json.ps1
@@ -121,6 +121,7 @@ function json_path([String] $json, [String] $jsonpath, [Hashtable] $substitution
 function json_path_legacy([String] $json, [String] $jsonpath, [Hashtable] $substitutions) {
     $result = $json | ConvertFrom-Json -ErrorAction Stop
     $isJsonPath = $jsonpath.StartsWith('$')
+
     debug $jsonpath
 
     foreach ($el in $jsonpath.split('.')) {